Login system (#61)
* Refactor the basic server architecture

I thought further about the basic server architecture and decided that
I had over-complicated the structure. So begin restructuring it.

* Fixed spelling mistake

* Relocated GlobalState to server package.

* Continued server refactoring

This change pushes the new use of GlobalState deeper into the server
code such that resource access is now completely converted to the
new pattern.

* Addressed comment

* Additional review comments

* Refactored staticServer

Aligned the staticServer with the code structure outlined in the
technical note.

* Minor addendum to the server design note.

* Renamed source file in server to better reflect purpose.

* Added support for cookie encryption keys.

Create cookie encryption keys. And format the code.

* Cookie decoding

This CL adds code to test incoming requests for the presence of an
auth cookie and to decode it if present.

* Pivot for wrapping architecture

After a false start, I decided to pivot on my architecture for chaining
of request handling modules. It will scale better and be harder to mess
up.

* Wrapped architecture pattern

This CL introduces the wrapped architecture for static resources. I
think by this point that I've settled on the server design pattern and
have the right avenues for extensibility. See server/doc.go for info
on the scheme.

* Start refactoring template code

Start refactoring the template code to use the wrap architecture.

* Converted template serving to wrap architecture.

* More cookie plumbing

This CL adds cookie access to the ResourceGuide entry server.

* Corrected formatting

* Added a display_name field to the password database.

* Added a display_name field to the cookie

* Add a login screen

Add a login screen and show login state on existing pages. Also, some
restructuring of how results are delivered to the template code.

* Added a basic login handler

Added a login handler that can receive the username and password combo.

* Interim check-in: login is tested.

Login flow tests and notes that one is not logged in. Code for actually
signing in will be added in a subsequent CL.

* Login flow supported

Now we have a basic login flow.

* Added a nicer display of login status.